Run add_is_active_to_brand migration only for public schema.
Former-commit-id: 168694a32249603df8a00e665a54a293375edbec [formerly 5bbfb8a26cf3926ab6a9acf0033d802db88c21b4] [formerly 11b61bfcbf879561f2f8f9ef46dd36f046cc1dc2 [formerly b46dffe0e0bf12394baa35545c6de2af510ead92]] [formerly 40d7b514b64217e4434bd924befa07baa8eae00d [formerly 6da213f955370b7b758da72abfda4139a4e3a918] [formerly c13a8574d8fcfb5c8099a25a397173c2923948c6 [formerly f811d9e2cc818a228709f3e2af314e6bfdce8116]]]
Former-commit-id: 86fc1886f0712aea02c9eb7c0c6193a803d35547 [formerly cc8c623a4c5a3f03754568897b023177156bf026] [formerly 05ddc003738c480cd9eab2fa997caac1e9a0de36 [formerly dcad083be91aca8ce13a70d5f18773c5777ae069]]
Former-commit-id: 246ee907c7f9761dd138a80cb4ebdb35d78ca07f [formerly afad30ac6acbc8d7d30a06ff14dd034447f27509]
Former-commit-id: d659324ae269758d692aff914c90b7c29c0469a9

+++ db/migrate/20150722130503_add_activity_indicator_to_brand.rb
@@ -1,9 +1,13 @@
 class AddActivityIndicatorToBrand < ActiveRecord::Migration
   def up
-    add_column :brands, :is_active, :boolean, default: true
+    if Apartment::Tenant.current == 'public'
+      add_column :brands, :is_active, :boolean, default: true
+    end
   end
 
   def down
-    remove_column :brands, :is_active
+    if Apartment::Tenant.current == 'public'
+      remove_column :brands, :is_active
+    end
   end
 end
